pyproject.toml
[project]
name = "youngwb"
version = "0.1.0"
description = "YoungWB using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==0.126.0",
    "crewai-tools>=0.9.1",
    "vnstock>=3.2.5",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
youngwb = "youngwb.main:run"
run_crew = "youngwb.main:run"
train = "youngwb.main:train"
replay = "youngwb.main:replay"
test = "youngwb.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
